refactor: PopupMenu 위치 수정
[#804]
